backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
mediapipe==0.10.7
Pillow==10.0.1
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13
playsound==1.3.0
scipy==1.11.2
werkzeug==2.3.7
eventlet==0.33.3
simple-websocket==0.10.1